Django-api/MP_django_dev/text2/urls.py
from django.urls import path, include
from django.conf import settings
from django.conf.urls import url
from django.conf.urls.static import static
from django.views.generic import RedirectView
from . import views


urlpatterns = [
    path('', views.graph)

]
