Django-api/MP_django_dev/requirements.txt
scipy
pandas
numpy
networkx
jgraph
djangorestframework
psycopg2